+++ src/Checkout/ExpressCheckout/SalesChannel/ExpressCreateOrderRoute.php
@@ -14,16 +14,21 @@ use Shopware\Core\Framework\Log\Package;
 use Shopware\Core\Framework\Plugin\Exception\DecorationPatternException;
 use Shopware\Core\Framework\Validation\DataBag\RequestDataBag;
 use Shopware\Core\System\SalesChannel\SalesChannelContext;
+use Shopware\Core\System\SystemConfig\SystemConfigService;
 use Swag\PayPal\Checkout\Cart\Service\CartPriceService;
 use Swag\PayPal\Checkout\Exception\OrderZeroValueException;
 use Swag\PayPal\Checkout\TokenResponse;
 use Swag\PayPal\OrdersApi\Builder\PayPalOrderBuilder;
 use Swag\PayPal\RestApi\PartnerAttributionId;
 use Swag\PayPal\RestApi\V2\Api\Order\ApplicationContext;
+use Swag\PayPal\RestApi\V2\Api\Order\PaymentSource\Common\Attributes\OrderUpdateCallbackConfig;
 use Swag\PayPal\RestApi\V2\Resource\OrderResource;
+use Swag\PayPal\Setting\Settings;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Attribute\Route;
+use Symfony\Component\Routing\Generator\UrlGeneratorInterface;
+use Symfony\Component\Routing\RouterInterface;
 
 #[Package('checkout')]
 #[Route(defaults: ['_routeScope' => ['store-api']])]
@@ -37,6 +42,8 @@ class ExpressCreateOrderRoute extends AbstractExpressCreateOrderRoute
         private readonly PayPalOrderBuilder $paypalOrderBuilder,
         private readonly OrderResource $orderResource,
         private readonly CartPriceService $cartPriceService,
+        private readonly SystemConfigService $systemConfigService,
+        private readonly RouterInterface $router,
         private readonly LoggerInterface $logger,
     ) {
     }
@@ -69,8 +76,29 @@ class ExpressCreateOrderRoute extends AbstractExpressCreateOrderRoute
 
             $this->logger->debug('Building order');
             $order = $this->paypalOrderBuilder->getOrderFromCart($cart, $salesChannelContext, new RequestDataBag($request->request->all()));
-            $order->getPaymentSource()?->getPaypal()?->getExperienceContext()->setShippingPreference(ApplicationContext::SHIPPING_PREFERENCE_GET_FROM_FILE);
-            $order->getPaymentSource()?->getPaypal()?->getExperienceContext()->setUserAction(ApplicationContext::USER_ACTION_CONTINUE);
+            $experienceContext = $order->getPaymentSource()?->getPaypal()?->getExperienceContext();
+
+            if ($experienceContext !== null) {
+                $experienceContext->setShippingPreference(ApplicationContext::SHIPPING_PREFERENCE_GET_FROM_FILE);
+                $experienceContext->setUserAction(ApplicationContext::USER_ACTION_CONTINUE);
+
+                // Configure shipping callback for dynamic price recalculation
+                if (!$this->systemConfigService->getBool(Settings::IS_LOCAL_ENVIRONMENT, $salesChannelContext->getSalesChannelId())) {
+                    $callbackConfig = new OrderUpdateCallbackConfig();
+                    $callbackUrl = $this->router->generate(
+                        'store-api.paypal.express.shipping_callback',
+                        ['salesChannelId' => $salesChannelContext->getSalesChannelId(), 'token' => $salesChannelContext->getToken()],
+                        UrlGeneratorInterface::ABSOLUTE_URL,
+                    );
+                    $callbackConfig->setCallbackUrl($callbackUrl);
+                    $callbackConfig->setCallbackEvents([OrderUpdateCallbackConfig::CALLBACK_EVENT_SHIPPING_OPTIONS]);
+                    $experienceContext->setOrderUpdateCallbackConfig($callbackConfig);
+
+                    $this->logger->debug('Configured shipping callback', ['callbackUrl' => $callbackUrl]);
+                } else {
+                    $this->logger->debug('Skipped shipping callback due to being disabled in system config');
+                }
+            }
 
             $orderResponse = $this->orderResource->create(
                 $order,