+++ src/Resources/app/administration/src/types/openapi.d.ts
@@ -1030,6 +1030,10 @@ export interface components {
     swag_paypal_v2_order_payment_source_common_attributes_customer: {
       id: string;
     };
+    swag_paypal_v2_order_payment_source_common_attributes_order_update_callback_config: {
+      callback_url: string;
+      callback_events: ("SHIPPING_ADDRESS" | "SHIPPING_OPTIONS")[];
+    };
     swag_paypal_v2_order_payment_source_common_attributes_vault: {
       id: string | null;
       store_in_vault: string;
@@ -1071,6 +1075,7 @@ export interface components {
       payment_method_preference: "UNRESTRICTED" | "IMMEDIATE_PAYMENT_REQUIRED";
       /** @description Only: PUI */
       customer_service_instructions: string[];
+      order_update_callback_config: components["schemas"]["swag_paypal_v2_order_payment_source_common_attributes_order_update_callback_config"];
     };
     swag_paypal_v2_order_payment_source_common_phone: {
       phone_type: string;
@@ -1189,6 +1194,7 @@ export interface components {
       items: components["schemas"]["swag_paypal_v2_order_purchase_unit_item"][] | null;
       shipping: components["schemas"]["swag_paypal_v2_order_purchase_unit_shipping"];
       payments: components["schemas"]["swag_paypal_v2_order_purchase_unit_payments"] | null;
+      shipping_options: components["schemas"]["swag_paypal_v2_order_purchase_unit_shipping_option"][] | null;
     };
     swag_paypal_v2_order_purchase_unit_amount: components["schemas"]["swag_paypal_v2_common_money"] & ({
       breakdown: components["schemas"]["swag_paypal_v2_order_purchase_unit_amount_breakdown"] | null;
@@ -1310,6 +1316,14 @@ export interface components {
       url: string | null;
       image_url: string | null;
     };
+    swag_paypal_v2_order_purchase_unit_shipping_option: {
+      id: string;
+      label: string;
+      amount: components["schemas"]["swag_paypal_v2_common_money"];
+      /** @enum {string} */
+      type: "SHIPPING" | "PICKUP";
+      selected: boolean;
+    };
     swag_paypal_v2_order_tracker: {
       capture_id: string;
       tracking_number: string;
@@ -1318,6 +1332,12 @@ export interface components {
       /** @default false */
       notify_payer: boolean;
       items: components["schemas"]["swag_paypal_v2_order_purchase_unit_shipping_tracker_item"][];
+    };
+    swag_paypal_v2_order_shipping_callback: {
+      id: string;
+      shipping_address: components["schemas"]["swag_paypal_v2_common_address"];
+      shipping_option: components["schemas"]["swag_paypal_v2_order_purchase_unit_shipping_option"];
+      purchase_units: components["schemas"]["swag_paypal_v2_order_purchase_unit"][];
     };
     swag_paypal_v2_patch: {
       op: string;