+++ src/Resources/app/administration/src/types/system-config.ts
@@ -75,6 +75,7 @@ export declare type SystemConfig = {
     'SwagPayPal.settings.crossBorderBuyerCountry'?: typeof COUNTRY_OVERRIDES[number] | null;
 
     'SwagPayPal.settings.agentCommerceOnboarded'?: string;
+    'SwagPayPal.settings.isLocalEnvironment'?: boolean;
 };
 
 /**
@@ -141,4 +142,6 @@ export const SystemConfigDefinition: Record<SYSTEM_CONFIG, 'string' | 'boolean' 
 
     'SwagPayPal.settings.crossBorderMessagingEnabled': 'boolean',
     'SwagPayPal.settings.crossBorderBuyerCountry': 'string',
+
+    'SwagPayPal.settings.isLocalEnvironment': 'boolean',
 };